get rid of LibTomCrypt derived hash implementations
