+++ src/Makefile.ktest.include
@@ -19,7 +19,8 @@ komodo_test_SOURCES = \
     test-komodo/test_netbase_tests.cpp \
     test-komodo/test_events.cpp \
     test-komodo/test_hex.cpp \
-    test-komodo/test_haraka_removal.cpp
+    test-komodo/test_haraka_removal.cpp \
+    test-komodo/test_oldhash_removal.cpp
 
 komodo_test_CPPFLAGS = $(komodod_CPPFLAGS)
