+++ src/komodo_utils.cpp
@@ -22,249 +22,8 @@
 
 #include "cc/CCinclude.h"
 
-// todo remove
-//struct komodo_state KOMODO_STATES[34];
-//int32_t ASSETCHAINS_CBMATURITY;
-//uint64_t ASSETCHAINS_TIMEUNLOCKFROM = 0;
-//uint64_t ASSETCHAINS_TIMEUNLOCKTO = 0;
-
-struct sha256_vstate { uint64_t length; uint32_t state[8],curlen; uint8_t buf[64]; };
-
-// following is ported from libtom
-
-#define STORE32L(x, y)                                                                     \
-{ (y)[3] = (uint8_t)(((x)>>24)&255); (y)[2] = (uint8_t)(((x)>>16)&255);   \
-(y)[1] = (uint8_t)(((x)>>8)&255); (y)[0] = (uint8_t)((x)&255); }
-
-#define LOAD32L(x, y)                            \
-{ x = (uint32_t)(((uint64_t)((y)[3] & 255)<<24) | \
-((uint32_t)((y)[2] & 255)<<16) | \
-((uint32_t)((y)[1] & 255)<<8)  | \
-((uint32_t)((y)[0] & 255))); }
-
-#define STORE64L(x, y)                                                                     \
-{ (y)[7] = (uint8_t)(((x)>>56)&255); (y)[6] = (uint8_t)(((x)>>48)&255);   \
-(y)[5] = (uint8_t)(((x)>>40)&255); (y)[4] = (uint8_t)(((x)>>32)&255);   \
-(y)[3] = (uint8_t)(((x)>>24)&255); (y)[2] = (uint8_t)(((x)>>16)&255);   \
-(y)[1] = (uint8_t)(((x)>>8)&255); (y)[0] = (uint8_t)((x)&255); }
-
-#define LOAD64L(x, y)                                                       \
-{ x = (((uint64_t)((y)[7] & 255))<<56)|(((uint64_t)((y)[6] & 255))<<48)| \
-(((uint64_t)((y)[5] & 255))<<40)|(((uint64_t)((y)[4] & 255))<<32)| \
-(((uint64_t)((y)[3] & 255))<<24)|(((uint64_t)((y)[2] & 255))<<16)| \
-(((uint64_t)((y)[1] & 255))<<8)|(((uint64_t)((y)[0] & 255))); }
-
-#define STORE32H(x, y)                                                                     \
-{ (y)[0] = (uint8_t)(((x)>>24)&255); (y)[1] = (uint8_t)(((x)>>16)&255);   \
-(y)[2] = (uint8_t)(((x)>>8)&255); (y)[3] = (uint8_t)((x)&255); }
-
-#define LOAD32H(x, y)                            \
-{ x = (uint32_t)(((uint64_t)((y)[0] & 255)<<24) | \
-((uint32_t)((y)[1] & 255)<<16) | \
-((uint32_t)((y)[2] & 255)<<8)  | \
-((uint32_t)((y)[3] & 255))); }
-
-#define STORE64H(x, y)                                                                     \
-{ (y)[0] = (uint8_t)(((x)>>56)&255); (y)[1] = (uint8_t)(((x)>>48)&255);     \
-(y)[2] = (uint8_t)(((x)>>40)&255); (y)[3] = (uint8_t)(((x)>>32)&255);     \
-(y)[4] = (uint8_t)(((x)>>24)&255); (y)[5] = (uint8_t)(((x)>>16)&255);     \
-(y)[6] = (uint8_t)(((x)>>8)&255); (y)[7] = (uint8_t)((x)&255); }
-
-#define LOAD64H(x, y)                                                      \
-{ x = (((uint64_t)((y)[0] & 255))<<56)|(((uint64_t)((y)[1] & 255))<<48) | \
-(((uint64_t)((y)[2] & 255))<<40)|(((uint64_t)((y)[3] & 255))<<32) | \
-(((uint64_t)((y)[4] & 255))<<24)|(((uint64_t)((y)[5] & 255))<<16) | \
-(((uint64_t)((y)[6] & 255))<<8)|(((uint64_t)((y)[7] & 255))); }
-
-// Various logical functions
-#define RORc(x, y) ( ((((uint32_t)(x)&0xFFFFFFFFUL)>>(uint32_t)((y)&31)) | ((uint32_t)(x)<<(uint32_t)(32-((y)&31)))) & 0xFFFFFFFFUL)
-#define Ch(x,y,z)       (z ^ (x & (y ^ z)))
-#define Maj(x,y,z)      (((x | y) & z) | (x & y))
-#define S(x, n)         RORc((x),(n))
-#define R(x, n)         (((x)&0xFFFFFFFFUL)>>(n))
-#define Sigma0(x)       (S(x, 2) ^ S(x, 13) ^ S(x, 22))
-#define Sigma1(x)       (S(x, 6) ^ S(x, 11) ^ S(x, 25))
-#define Gamma0(x)       (S(x, 7) ^ S(x, 18) ^ R(x, 3))
-#define Gamma1(x)       (S(x, 17) ^ S(x, 19) ^ R(x, 10))
-#define MIN(x, y) ( ((x)<(y))?(x):(y) )
-
-static inline int32_t sha256_vcompress(struct sha256_vstate * md,uint8_t *buf)
-{
-    uint32_t S[8],W[64],t0,t1,i;
-    for (i=0; i<8; i++) // copy state into S
-        S[i] = md->state[i];
-    for (i=0; i<16; i++) // copy the state into 512-bits into W[0..15]
-        LOAD32H(W[i],buf + (4*i));
-    for (i=16; i<64; i++) // fill W[16..63]
-        W[i] = Gamma1(W[i - 2]) + W[i - 7] + Gamma0(W[i - 15]) + W[i - 16];
-
-#define RND(a,b,c,d,e,f,g,h,i,ki)                    \
-t0 = h + Sigma1(e) + Ch(e, f, g) + ki + W[i];   \
-t1 = Sigma0(a) + Maj(a, b, c);                  \
-d += t0;                                        \
-h  = t0 + t1;
-
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],0,0x428a2f98);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],1,0x71374491);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],2,0xb5c0fbcf);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],3,0xe9b5dba5);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],4,0x3956c25b);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],5,0x59f111f1);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],6,0x923f82a4);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],7,0xab1c5ed5);
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],8,0xd807aa98);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],9,0x12835b01);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],10,0x243185be);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],11,0x550c7dc3);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],12,0x72be5d74);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],13,0x80deb1fe);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],14,0x9bdc06a7);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],15,0xc19bf174);
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],16,0xe49b69c1);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],17,0xefbe4786);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],18,0x0fc19dc6);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],19,0x240ca1cc);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],20,0x2de92c6f);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],21,0x4a7484aa);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],22,0x5cb0a9dc);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],23,0x76f988da);
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],24,0x983e5152);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],25,0xa831c66d);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],26,0xb00327c8);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],27,0xbf597fc7);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],28,0xc6e00bf3);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],29,0xd5a79147);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],30,0x06ca6351);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],31,0x14292967);
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],32,0x27b70a85);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],33,0x2e1b2138);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],34,0x4d2c6dfc);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],35,0x53380d13);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],36,0x650a7354);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],37,0x766a0abb);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],38,0x81c2c92e);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],39,0x92722c85);
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],40,0xa2bfe8a1);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],41,0xa81a664b);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],42,0xc24b8b70);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],43,0xc76c51a3);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],44,0xd192e819);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],45,0xd6990624);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],46,0xf40e3585);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],47,0x106aa070);
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],48,0x19a4c116);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],49,0x1e376c08);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],50,0x2748774c);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],51,0x34b0bcb5);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],52,0x391c0cb3);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],53,0x4ed8aa4a);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],54,0x5b9cca4f);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],55,0x682e6ff3);
-    RND(S[0],S[1],S[2],S[3],S[4],S[5],S[6],S[7],56,0x748f82ee);
-    RND(S[7],S[0],S[1],S[2],S[3],S[4],S[5],S[6],57,0x78a5636f);
-    RND(S[6],S[7],S[0],S[1],S[2],S[3],S[4],S[5],58,0x84c87814);
-    RND(S[5],S[6],S[7],S[0],S[1],S[2],S[3],S[4],59,0x8cc70208);
-    RND(S[4],S[5],S[6],S[7],S[0],S[1],S[2],S[3],60,0x90befffa);
-    RND(S[3],S[4],S[5],S[6],S[7],S[0],S[1],S[2],61,0xa4506ceb);
-    RND(S[2],S[3],S[4],S[5],S[6],S[7],S[0],S[1],62,0xbef9a3f7);
-    RND(S[1],S[2],S[3],S[4],S[5],S[6],S[7],S[0],63,0xc67178f2);
-#undef RND
-    for (i=0; i<8; i++) // feedback
-        md->state[i] = md->state[i] + S[i];
-    return(0);
-}
-
-#undef RORc
-#undef Ch
-#undef Maj
-#undef S
-#undef R
-#undef Sigma0
-#undef Sigma1
-#undef Gamma0
-#undef Gamma1
-
-static inline int32_t sha256_vdone(struct sha256_vstate *md,uint8_t *out)
-{
-    int32_t i;
-    if ( md->curlen >= sizeof(md->buf) )
-        return(-1);
-    md->length += md->curlen * 8; // increase the length of the message
-    md->buf[md->curlen++] = (uint8_t)0x80; // append the '1' bit
-    // if len > 56 bytes we append zeros then compress.  Then we can fall back to padding zeros and length encoding like normal.
-    if ( md->curlen > 56 )
-    {
-        while ( md->curlen < 64 )
-            md->buf[md->curlen++] = (uint8_t)0;
-        sha256_vcompress(md,md->buf);
-        md->curlen = 0;
-    }
-    while ( md->curlen < 56 ) // pad upto 56 bytes of zeroes
-        md->buf[md->curlen++] = (uint8_t)0;
-    STORE64H(md->length,md->buf+56); // store length
-    sha256_vcompress(md,md->buf);
-    for (i=0; i<8; i++) // copy output
-        STORE32H(md->state[i],out+(4*i));
-    return(0);
-}
-
-static inline int32_t sha256_vprocess(struct sha256_vstate *md,const uint8_t *in,uint64_t inlen)
-{
-    uint64_t n; int32_t err;
-    if ( md->curlen > sizeof(md->buf) )
-        return(-1);
-    while ( inlen > 0 )
-    {
-        if ( md->curlen == 0 && inlen >= 64 )
-        {
-            if ( (err= sha256_vcompress(md,(uint8_t *)in)) != 0 )
-                return(err);
-            md->length += 64 * 8, in += 64, inlen -= 64;
-        }
-        else
-        {
-            n = MIN(inlen,64 - md->curlen);
-            memcpy(md->buf + md->curlen,in,(size_t)n);
-            md->curlen += n, in += n, inlen -= n;
-            if ( md->curlen == 64 )
-            {
-                if ( (err= sha256_vcompress(md,md->buf)) != 0 )
-                    return(err);
-                md->length += 8*64;
-                md->curlen = 0;
-            }
-        }
-    }
-    return(0);
-}
-
-static inline void sha256_vinit(struct sha256_vstate * md)
-{
-    md->curlen = 0;
-    md->length = 0;
-    md->state[0] = 0x6A09E667UL;
-    md->state[1] = 0xBB67AE85UL;
-    md->state[2] = 0x3C6EF372UL;
-    md->state[3] = 0xA54FF53AUL;
-    md->state[4] = 0x510E527FUL;
-    md->state[5] = 0x9B05688CUL;
-    md->state[6] = 0x1F83D9ABUL;
-    md->state[7] = 0x5BE0CD19UL;
-}
-
 void vcalc_sha256(char deprecated[(256 >> 3) * 2 + 1],uint8_t hash[256 >> 3],uint8_t *src,int32_t len)
 {
-    /* 
-    struct sha256_vstate md;
-    sha256_vinit(&md);
-    sha256_vprocess(&md,src,len);
-    sha256_vdone(&md,hash);
-    */
-
-    // we will use CSHA256 class instead of above implementation,
-    // in case if daemon compiled with USE_ASM enabled it will use
-    // hardware (SSE4) implementation, otherwise standart
-
     CSHA256().Write((const unsigned char *)src, len).Finalize(hash);
 }
 
@@ -277,397 +36,6 @@ bits256 bits256_doublesha256(char *deprecated,uint8_t *data,int32_t datalen)
         hash.bytes[i] = hash2.bytes[sizeof(hash) - 1 - i];
     return(hash);
 }
-
-// rmd160: the five basic functions F(), G() and H()
-#define F(x, y, z)        ((x) ^ (y) ^ (z))
-#define G(x, y, z)        (((x) & (y)) | (~(x) & (z)))
-#define H(x, y, z)        (((x) | ~(y)) ^ (z))
-#define I(x, y, z)        (((x) & (z)) | ((y) & ~(z)))
-#define J(x, y, z)        ((x) ^ ((y) | ~(z)))
-#define ROLc(x, y) ( (((unsigned long)(x)<<(unsigned long)((y)&31)) | (((unsigned long)(x)&0xFFFFFFFFUL)>>(unsigned long)(32-((y)&31)))) & 0xFFFFFFFFUL)
-
-/* the ten basic operations FF() through III() */
-#define FF(a, b, c, d, e, x, s)        \
-(a) += F((b), (c), (d)) + (x);\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define GG(a, b, c, d, e, x, s)        \
-(a) += G((b), (c), (d)) + (x) + 0x5a827999UL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define HH(a, b, c, d, e, x, s)        \
-(a) += H((b), (c), (d)) + (x) + 0x6ed9eba1UL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define II(a, b, c, d, e, x, s)        \
-(a) += I((b), (c), (d)) + (x) + 0x8f1bbcdcUL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define JJ(a, b, c, d, e, x, s)        \
-(a) += J((b), (c), (d)) + (x) + 0xa953fd4eUL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define FFF(a, b, c, d, e, x, s)        \
-(a) += F((b), (c), (d)) + (x);\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define GGG(a, b, c, d, e, x, s)        \
-(a) += G((b), (c), (d)) + (x) + 0x7a6d76e9UL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define HHH(a, b, c, d, e, x, s)        \
-(a) += H((b), (c), (d)) + (x) + 0x6d703ef3UL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define III(a, b, c, d, e, x, s)        \
-(a) += I((b), (c), (d)) + (x) + 0x5c4dd124UL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-#define JJJ(a, b, c, d, e, x, s)        \
-(a) += J((b), (c), (d)) + (x) + 0x50a28be6UL;\
-(a) = ROLc((a), (s)) + (e);\
-(c) = ROLc((c), 10);
-
-static int32_t rmd160_vcompress(struct rmd160_vstate *md,uint8_t *buf)
-{
-    uint32_t aa,bb,cc,dd,ee,aaa,bbb,ccc,ddd,eee,X[16];
-    int i;
-
-    /* load words X */
-    for (i = 0; i < 16; i++){
-        LOAD32L(X[i], buf + (4 * i));
-    }
-
-    /* load state */
-    aa = aaa = md->state[0];
-    bb = bbb = md->state[1];
-    cc = ccc = md->state[2];
-    dd = ddd = md->state[3];
-    ee = eee = md->state[4];
-
-    /* round 1 */
-    FF(aa, bb, cc, dd, ee, X[ 0], 11);
-    FF(ee, aa, bb, cc, dd, X[ 1], 14);
-    FF(dd, ee, aa, bb, cc, X[ 2], 15);
-    FF(cc, dd, ee, aa, bb, X[ 3], 12);
-    FF(bb, cc, dd, ee, aa, X[ 4],  5);
-    FF(aa, bb, cc, dd, ee, X[ 5],  8);
-    FF(ee, aa, bb, cc, dd, X[ 6],  7);
-    FF(dd, ee, aa, bb, cc, X[ 7],  9);
-    FF(cc, dd, ee, aa, bb, X[ 8], 11);
-    FF(bb, cc, dd, ee, aa, X[ 9], 13);
-    FF(aa, bb, cc, dd, ee, X[10], 14);
-    FF(ee, aa, bb, cc, dd, X[11], 15);
-    FF(dd, ee, aa, bb, cc, X[12],  6);
-    FF(cc, dd, ee, aa, bb, X[13],  7);
-    FF(bb, cc, dd, ee, aa, X[14],  9);
-    FF(aa, bb, cc, dd, ee, X[15],  8);
-
-    /* round 2 */
-    GG(ee, aa, bb, cc, dd, X[ 7],  7);
-    GG(dd, ee, aa, bb, cc, X[ 4],  6);
-    GG(cc, dd, ee, aa, bb, X[13],  8);
-    GG(bb, cc, dd, ee, aa, X[ 1], 13);
-    GG(aa, bb, cc, dd, ee, X[10], 11);
-    GG(ee, aa, bb, cc, dd, X[ 6],  9);
-    GG(dd, ee, aa, bb, cc, X[15],  7);
-    GG(cc, dd, ee, aa, bb, X[ 3], 15);
-    GG(bb, cc, dd, ee, aa, X[12],  7);
-    GG(aa, bb, cc, dd, ee, X[ 0], 12);
-    GG(ee, aa, bb, cc, dd, X[ 9], 15);
-    GG(dd, ee, aa, bb, cc, X[ 5],  9);
-    GG(cc, dd, ee, aa, bb, X[ 2], 11);
-    GG(bb, cc, dd, ee, aa, X[14],  7);
-    GG(aa, bb, cc, dd, ee, X[11], 13);
-    GG(ee, aa, bb, cc, dd, X[ 8], 12);
-
-    /* round 3 */
-    HH(dd, ee, aa, bb, cc, X[ 3], 11);
-    HH(cc, dd, ee, aa, bb, X[10], 13);
-    HH(bb, cc, dd, ee, aa, X[14],  6);
-    HH(aa, bb, cc, dd, ee, X[ 4],  7);
-    HH(ee, aa, bb, cc, dd, X[ 9], 14);
-    HH(dd, ee, aa, bb, cc, X[15],  9);
-    HH(cc, dd, ee, aa, bb, X[ 8], 13);
-    HH(bb, cc, dd, ee, aa, X[ 1], 15);
-    HH(aa, bb, cc, dd, ee, X[ 2], 14);
-    HH(ee, aa, bb, cc, dd, X[ 7],  8);
-    HH(dd, ee, aa, bb, cc, X[ 0], 13);
-    HH(cc, dd, ee, aa, bb, X[ 6],  6);
-    HH(bb, cc, dd, ee, aa, X[13],  5);
-    HH(aa, bb, cc, dd, ee, X[11], 12);
-    HH(ee, aa, bb, cc, dd, X[ 5],  7);
-    HH(dd, ee, aa, bb, cc, X[12],  5);
-
-    /* round 4 */
-    II(cc, dd, ee, aa, bb, X[ 1], 11);
-    II(bb, cc, dd, ee, aa, X[ 9], 12);
-    II(aa, bb, cc, dd, ee, X[11], 14);
-    II(ee, aa, bb, cc, dd, X[10], 15);
-    II(dd, ee, aa, bb, cc, X[ 0], 14);
-    II(cc, dd, ee, aa, bb, X[ 8], 15);
-    II(bb, cc, dd, ee, aa, X[12],  9);
-    II(aa, bb, cc, dd, ee, X[ 4],  8);
-    II(ee, aa, bb, cc, dd, X[13],  9);
-    II(dd, ee, aa, bb, cc, X[ 3], 14);
-    II(cc, dd, ee, aa, bb, X[ 7],  5);
-    II(bb, cc, dd, ee, aa, X[15],  6);
-    II(aa, bb, cc, dd, ee, X[14],  8);
-    II(ee, aa, bb, cc, dd, X[ 5],  6);
-    II(dd, ee, aa, bb, cc, X[ 6],  5);
-    II(cc, dd, ee, aa, bb, X[ 2], 12);
-
-    /* round 5 */
-    JJ(bb, cc, dd, ee, aa, X[ 4],  9);
-    JJ(aa, bb, cc, dd, ee, X[ 0], 15);
-    JJ(ee, aa, bb, cc, dd, X[ 5],  5);
-    JJ(dd, ee, aa, bb, cc, X[ 9], 11);
-    JJ(cc, dd, ee, aa, bb, X[ 7],  6);
-    JJ(bb, cc, dd, ee, aa, X[12],  8);
-    JJ(aa, bb, cc, dd, ee, X[ 2], 13);
-    JJ(ee, aa, bb, cc, dd, X[10], 12);
-    JJ(dd, ee, aa, bb, cc, X[14],  5);
-    JJ(cc, dd, ee, aa, bb, X[ 1], 12);
-    JJ(bb, cc, dd, ee, aa, X[ 3], 13);
-    JJ(aa, bb, cc, dd, ee, X[ 8], 14);
-    JJ(ee, aa, bb, cc, dd, X[11], 11);
-    JJ(dd, ee, aa, bb, cc, X[ 6],  8);
-    JJ(cc, dd, ee, aa, bb, X[15],  5);
-    JJ(bb, cc, dd, ee, aa, X[13],  6);
-
-    /* parallel round 1 */
-    JJJ(aaa, bbb, ccc, ddd, eee, X[ 5],  8);
-    JJJ(eee, aaa, bbb, ccc, ddd, X[14],  9);
-    JJJ(ddd, eee, aaa, bbb, ccc, X[ 7],  9);
-    JJJ(ccc, ddd, eee, aaa, bbb, X[ 0], 11);
-    JJJ(bbb, ccc, ddd, eee, aaa, X[ 9], 13);
-    JJJ(aaa, bbb, ccc, ddd, eee, X[ 2], 15);
-    JJJ(eee, aaa, bbb, ccc, ddd, X[11], 15);
-    JJJ(ddd, eee, aaa, bbb, ccc, X[ 4],  5);
-    JJJ(ccc, ddd, eee, aaa, bbb, X[13],  7);
-    JJJ(bbb, ccc, ddd, eee, aaa, X[ 6],  7);
-    JJJ(aaa, bbb, ccc, ddd, eee, X[15],  8);
-    JJJ(eee, aaa, bbb, ccc, ddd, X[ 8], 11);
-    JJJ(ddd, eee, aaa, bbb, ccc, X[ 1], 14);
-    JJJ(ccc, ddd, eee, aaa, bbb, X[10], 14);
-    JJJ(bbb, ccc, ddd, eee, aaa, X[ 3], 12);
-    JJJ(aaa, bbb, ccc, ddd, eee, X[12],  6);
-
-    /* parallel round 2 */
-    III(eee, aaa, bbb, ccc, ddd, X[ 6],  9);
-    III(ddd, eee, aaa, bbb, ccc, X[11], 13);
-    III(ccc, ddd, eee, aaa, bbb, X[ 3], 15);
-    III(bbb, ccc, ddd, eee, aaa, X[ 7],  7);
-    III(aaa, bbb, ccc, ddd, eee, X[ 0], 12);
-    III(eee, aaa, bbb, ccc, ddd, X[13],  8);
-    III(ddd, eee, aaa, bbb, ccc, X[ 5],  9);
-    III(ccc, ddd, eee, aaa, bbb, X[10], 11);
-    III(bbb, ccc, ddd, eee, aaa, X[14],  7);
-    III(aaa, bbb, ccc, ddd, eee, X[15],  7);
-    III(eee, aaa, bbb, ccc, ddd, X[ 8], 12);
-    III(ddd, eee, aaa, bbb, ccc, X[12],  7);
-    III(ccc, ddd, eee, aaa, bbb, X[ 4],  6);
-    III(bbb, ccc, ddd, eee, aaa, X[ 9], 15);
-    III(aaa, bbb, ccc, ddd, eee, X[ 1], 13);
-    III(eee, aaa, bbb, ccc, ddd, X[ 2], 11);
-
-    /* parallel round 3 */
-    HHH(ddd, eee, aaa, bbb, ccc, X[15],  9);
-    HHH(ccc, ddd, eee, aaa, bbb, X[ 5],  7);
-    HHH(bbb, ccc, ddd, eee, aaa, X[ 1], 15);
-    HHH(aaa, bbb, ccc, ddd, eee, X[ 3], 11);
-    HHH(eee, aaa, bbb, ccc, ddd, X[ 7],  8);
-    HHH(ddd, eee, aaa, bbb, ccc, X[14],  6);
-    HHH(ccc, ddd, eee, aaa, bbb, X[ 6],  6);
-    HHH(bbb, ccc, ddd, eee, aaa, X[ 9], 14);
-    HHH(aaa, bbb, ccc, ddd, eee, X[11], 12);
-    HHH(eee, aaa, bbb, ccc, ddd, X[ 8], 13);
-    HHH(ddd, eee, aaa, bbb, ccc, X[12],  5);
-    HHH(ccc, ddd, eee, aaa, bbb, X[ 2], 14);
-    HHH(bbb, ccc, ddd, eee, aaa, X[10], 13);
-    HHH(aaa, bbb, ccc, ddd, eee, X[ 0], 13);
-    HHH(eee, aaa, bbb, ccc, ddd, X[ 4],  7);
-    HHH(ddd, eee, aaa, bbb, ccc, X[13],  5);
-
-    /* parallel round 4 */
-    GGG(ccc, ddd, eee, aaa, bbb, X[ 8], 15);
-    GGG(bbb, ccc, ddd, eee, aaa, X[ 6],  5);
-    GGG(aaa, bbb, ccc, ddd, eee, X[ 4],  8);
-    GGG(eee, aaa, bbb, ccc, ddd, X[ 1], 11);
-    GGG(ddd, eee, aaa, bbb, ccc, X[ 3], 14);
-    GGG(ccc, ddd, eee, aaa, bbb, X[11], 14);
-    GGG(bbb, ccc, ddd, eee, aaa, X[15],  6);
-    GGG(aaa, bbb, ccc, ddd, eee, X[ 0], 14);
-    GGG(eee, aaa, bbb, ccc, ddd, X[ 5],  6);
-    GGG(ddd, eee, aaa, bbb, ccc, X[12],  9);
-    GGG(ccc, ddd, eee, aaa, bbb, X[ 2], 12);
-    GGG(bbb, ccc, ddd, eee, aaa, X[13],  9);
-    GGG(aaa, bbb, ccc, ddd, eee, X[ 9], 12);
-    GGG(eee, aaa, bbb, ccc, ddd, X[ 7],  5);
-    GGG(ddd, eee, aaa, bbb, ccc, X[10], 15);
-    GGG(ccc, ddd, eee, aaa, bbb, X[14],  8);
-
-    /* parallel round 5 */
-    FFF(bbb, ccc, ddd, eee, aaa, X[12] ,  8);
-    FFF(aaa, bbb, ccc, ddd, eee, X[15] ,  5);
-    FFF(eee, aaa, bbb, ccc, ddd, X[10] , 12);
-    FFF(ddd, eee, aaa, bbb, ccc, X[ 4] ,  9);
-    FFF(ccc, ddd, eee, aaa, bbb, X[ 1] , 12);
-    FFF(bbb, ccc, ddd, eee, aaa, X[ 5] ,  5);
-    FFF(aaa, bbb, ccc, ddd, eee, X[ 8] , 14);
-    FFF(eee, aaa, bbb, ccc, ddd, X[ 7] ,  6);
-    FFF(ddd, eee, aaa, bbb, ccc, X[ 6] ,  8);
-    FFF(ccc, ddd, eee, aaa, bbb, X[ 2] , 13);
-    FFF(bbb, ccc, ddd, eee, aaa, X[13] ,  6);
-    FFF(aaa, bbb, ccc, ddd, eee, X[14] ,  5);
-    FFF(eee, aaa, bbb, ccc, ddd, X[ 0] , 15);
-    FFF(ddd, eee, aaa, bbb, ccc, X[ 3] , 13);
-    FFF(ccc, ddd, eee, aaa, bbb, X[ 9] , 11);
-    FFF(bbb, ccc, ddd, eee, aaa, X[11] , 11);
-
-    /* combine results */
-    ddd += cc + md->state[1];               /* final result for md->state[0] */
-    md->state[1] = md->state[2] + dd + eee;
-    md->state[2] = md->state[3] + ee + aaa;
-    md->state[3] = md->state[4] + aa + bbb;
-    md->state[4] = md->state[0] + bb + ccc;
-    md->state[0] = ddd;
-
-    return 0;
-}
-
-/**
- Initialize the hash state
- @param md   The hash state you wish to initialize
- @return 0 if successful
- */
-int rmd160_vinit(struct rmd160_vstate * md)
-{
-    md->state[0] = 0x67452301UL;
-    md->state[1] = 0xefcdab89UL;
-    md->state[2] = 0x98badcfeUL;
-    md->state[3] = 0x10325476UL;
-    md->state[4] = 0xc3d2e1f0UL;
-    md->curlen   = 0;
-    md->length   = 0;
-    return 0;
-}
-
-/**
- Process a block of memory though the hash
- @param md     The hash state
- @param in     The data to hash
- @param inlen  The length of the data (octets)
- @return 0 if successful
- */
-int rmd160_vprocess (struct rmd160_vstate * md, const unsigned char *in, unsigned long inlen)               
-{
-    unsigned long n;
-    int           err;
-    if (md->curlen > sizeof(md->buf)) {
-        return -1;
-    }
-    while (inlen > 0) {
-        if (md->curlen == 0 && inlen >= 64) {
-            if ((err = rmd160_vcompress (md, (unsigned char *)in)) != 0) {
-                return err;
-            }
-            md->length += 64 * 8;
-            in             += 64;
-            inlen          -= 64;
-        } else {
-            n = MIN(inlen, (64 - md->curlen));
-            memcpy(md->buf + md->curlen, in, (size_t)n);
-            md->curlen += n;
-            in             += n;
-            inlen          -= n;
-            if (md->curlen == 64) {
-                if ((err = rmd160_vcompress (md, md->buf)) != 0) {
-                    return err;
-                }
-                md->length += 8*64;
-                md->curlen = 0;
-            }
-        }
-    }
-    return 0;
-}
-
-/**
- Terminate the hash to get the digest
- @param md  The hash state
- @param out [out] The destination of the hash (20 bytes)
- @return 0 if successful
- */
-int rmd160_vdone(struct rmd160_vstate * md, unsigned char *out)
-{
-    int i;
-    if (md->curlen >= sizeof(md->buf)) {
-        return -1;
-    }
-    /* increase the length of the message */
-    md->length += md->curlen * 8;
-
-    /* append the '1' bit */
-    md->buf[md->curlen++] = (unsigned char)0x80;
-
-    /* if the length is currently above 56 bytes we append zeros
-     * then compress.  Then we can fall back to padding zeros and length
-     * encoding like normal.
-     */
-    if (md->curlen > 56) {
-        while (md->curlen < 64) {
-            md->buf[md->curlen++] = (unsigned char)0;
-        }
-        rmd160_vcompress(md, md->buf);
-        md->curlen = 0;
-    }
-    /* pad upto 56 bytes of zeroes */
-    while (md->curlen < 56) {
-        md->buf[md->curlen++] = (unsigned char)0;
-    }
-    /* store length */
-    STORE64L(md->length, md->buf+56);
-    rmd160_vcompress(md, md->buf);
-    /* copy output */
-    for (i = 0; i < 5; i++) {
-        STORE32L(md->state[i], out+(4*i));
-    }
-    return 0;
-}
-
-void calc_rmd160(char deprecated[41],uint8_t buf[20],uint8_t *msg,int32_t len)
-{
-    struct rmd160_vstate md;
-    rmd160_vinit(&md);
-    rmd160_vprocess(&md,msg,len);
-    rmd160_vdone(&md, buf);
-}
-#undef F
-#undef G
-#undef H
-#undef I
-#undef J
-#undef ROLc
-#undef FF
-#undef GG
-#undef HH
-#undef II
-#undef JJ
-#undef FFF
-#undef GGG
-#undef HHH
-#undef III
-#undef JJJ
 
 static const uint32_t crc32_tab[] = {
 	0x00000000, 0x77073096, 0xee0e612c, 0x990951ba, 0x076dc419, 0x706af48f,
@@ -730,9 +98,7 @@ uint32_t calc_crc32(uint32_t crc,const void *buf,size_t size)
 
 void calc_rmd160_sha256(uint8_t rmd160[20],uint8_t *data,int32_t datalen)
 {
-    bits256 hash;
-    vcalc_sha256(0,hash.bytes,data,datalen);
-    calc_rmd160(0,rmd160,hash.bytes,sizeof(hash));
+    CHash160().Write((const unsigned char *)data, datalen).Finalize(rmd160); // SHA-256 + RIPEMD-160
 }
 
 int32_t bitcoin_addr2rmd160(uint8_t *addrtypep,uint8_t rmd160[20],const char *coinaddr)
